devops/metrics/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.2
awswrangler==3.4.2
blinker==1.7.0
boto3==1.34.14
botocore==1.34.14
bson==0.5.10
build==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses==0.6
dill==0.3.7
dnspython==2.4.2
fastapi==0.109.0
Flask==3.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pandas-stubs==2.1.4.231227
pip-tools==7.3.0
platformdirs==4.1.0
pluggy==1.3.0
pyarrow==14.0.2
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.0.3
pymongo==4.6.1
pyproject_hooks==1.0.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-mock==1.11.0
s3transfer==0.10.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
starlette==0.35.1
tomlkit==0.12.3
types-pytz==2023.3.1.1
types-requests==2.31.0.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
uvicorn==0.25.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
